feat: migrate Python backend to Rust-only architecture
- Remove entire Python server/ directory (moved to 项目废弃的python/)
- Add Rust sandbox executor for strategy execution
- Add health check endpoint
- Add market data module
- Update docker-compose for Rust-only deployment
- Add Dockerfile.local for local binary deployment

+++ server-rust/src/backtest/analyzer.rs
@@ -141,7 +141,7 @@ pub fn calculate_performance(
         total_trades: broker.trades.len() as u64,
         total_commission,
         final_value: final_val,
-        duration_days: duration.max(0) as u64,
+        duration_days: duration.max(0),
         trading_days,
         monthly_returns,
     }
